Generalise interface of possible_next_states().
The idea here is to make it possible for possible_next_states() to
choose to build a robot in several minutes (e.g. because it needs more
ore than there is on hand now).  This would allow is to choose an
action spanning a number of minutes, which should cut the search space
down.