Day 19: much faster, but wrong.
We abandoned the priority queue as it was memory-inefficient.   This
made it possible for us to complete the example test, but the outcome
of that test is that the answer is wrong.

The other tests which we'd always been running before still pass, so
although we know this version is broken, we don't know it's any more
broken than previous versions.